requirements.txt
datasets>=2.20.0
pandas>=2.2.0
numpy>=1.26.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
transformers>=4.44.0
accelerate>=0.33.0
torch>=2.3.0
pyarrow>=16.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
python-docx>=1.1.2
